LayoutTests/fast/dom/timer-clear-interval-in-handler-and-generate-error-expected.txt

CONSOLE MESSAGE: line 10: NOT_FOUND_ERR: DOM Exception 8: An attempt was made to reference a Node in a context where it does not exist.
